Add files via upload
This is a simple front-end implementation of the scoreboard described in the second option for the technical challenge. The implementation file "jquery-2.1.1.js" must be placed in the same directory as the other two files in order to run the jQuery defined in "ScoreTableJQ.js". The application allows any number of new players to be inserted, but only displays the top 20 at any given time, organized into two descending columns of 10.